extract to parser.ts
テストしやすいように。svelteコンポーネントをインポートしているファイルをjestでテストしようとしたら解決が難しいエラーが発生したため、テストしやすいようにファイルに切り出す。
